runtime: split landlock functionality out of demo main
