+++ runtime/landlock.c
@@ -1,28 +1,13 @@
+#include "landlock.h"
 #include "forbid_paths.h"
 #include "landlock_syscall.h"
-#include "strv.h"
 
 #include <errno.h>
-#include <fcntl.h>
 #include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
 
-int main(const int argc, char *const argv[], char *const *const envp) {
-  const char *cmd_path;
-  char *const *cmd_argv;
+int landlock_setup(struct landlock_ctx *ctx) {
   int ruleset_fd, abi_ver;
-  __u64 access_fs_rw = ACCESS_FS_ROUGHLY_READ | ACCESS_FS_ROUGHLY_WRITE;
-  struct landlock_ruleset_attr ruleset_attr = {
-      .handled_access_fs = access_fs_rw,
-  };
-
-  if (argc < 2) {
-    fprintf(stderr, "usage: DENY_PATH=... %s <command>\n\n", basename(argv[0]));
-    fprintf(stderr, "built against landlock ABI version <= %d\n",
-            LANDLOCK_ABI_LAST);
-    return 1;
-  }
+  ctx->ruleset_attr.handled_access_fs = ACCESS_FS_ROUGHLY_READ | ACCESS_FS_ROUGHLY_WRITE;
 
   abi_ver = landlock_create_ruleset(NULL, 0, LANDLOCK_CREATE_RULESET_VERSION);
   if (abi_ver < 0) {
@@ -67,11 +52,11 @@ int main(const int argc, char *const argv[], char *const *const envp) {
      * should then fall back to not restrict themselves at all if
      * the running kernel only supports ABI 1.
      */
-    ruleset_attr.handled_access_fs &= ~LANDLOCK_ACCESS_FS_REFER_OR_0;
+    ctx->ruleset_attr.handled_access_fs &= ~LANDLOCK_ACCESS_FS_REFER_OR_0;
     __attribute__((fallthrough));
   case 2:
     /* Removes LANDLOCK_ACCESS_FS_TRUNCATE for ABI < 3 */
-    ruleset_attr.handled_access_fs &= ~LANDLOCK_ACCESS_FS_TRUNCATE_OR_0;
+    ctx->ruleset_attr.handled_access_fs &= ~LANDLOCK_ACCESS_FS_TRUNCATE_OR_0;
 
     fprintf(stderr,
             "Hint: You should update the running kernel "
@@ -87,15 +72,16 @@ int main(const int argc, char *const argv[], char *const *const envp) {
         "rebuild sandboxer to use features from ABI version %d instead of %d\n",
         abi_ver, LANDLOCK_ABI_LAST);
   }
-  access_fs_rw &= ruleset_attr.handled_access_fs;
+  return 0;
+}
 
-  ruleset_fd = landlock_create_ruleset(&ruleset_attr, sizeof(ruleset_attr), 0);
+int landlock_apply_path_blacklist(const struct landlock_ctx *ctx, const char **paths) {
+  int ruleset_fd = landlock_create_ruleset(&ctx->ruleset_attr, sizeof(ctx->ruleset_attr), 0);
   if (ruleset_fd < 0) {
     perror("Failed to create ruleset");
     return 1;
   }
-  const char *path = getenv("DENY_PATH");
-  const char *paths[] = {path, NULL};
+
   if (forbid_paths(paths, ruleset_fd) < 0) {
     fprintf(stderr, "Failed to set up path allowlist\n");
     return 1;
@@ -110,12 +96,7 @@ int main(const int argc, char *const argv[], char *const *const envp) {
     goto err_close_ruleset;
   }
   close(ruleset_fd);
-
-  cmd_path = argv[1];
-  cmd_argv = argv + 1;
-  execvpe(cmd_path, cmd_argv, envp);
-  perror("execvpe");
-  return 1;
+  return 0;
 
 err_close_ruleset:
   close(ruleset_fd);
